クラスごとにファイルを分けた。余計なコメントを消した。
